made "Table" component a ProjectsOverview view since it contains all the components
